chore: split the components
