+++ src/features/packageManager/components/index.tsx
@@ -2,3 +2,4 @@ export * from "./PackageManager";
 export * from "./PackageManagerSearch";
 export * from "./EnvironmentDropdown";
 export * from "./Environment";
+export * from "./EnvironmentsList";
